got entire stores object to display
